Fixed issue #13
`.scss` files were not being compiled in the dev build. The real case
of the problem was that `dist\tmp` is being shared between dev and prod
builds. This directory was not being cleaned up during a dev build. So
if there were files sitting in `dist\temp` from the prod build, then
dev build did not like it. Fixed by cleaning up tmp directory in during
dev builds.